Unison.Server.Backend

Types

data BackendError

Constructors

NoSuchNamespace Absolute

BadNamespace

Fields

String

error message

String

namespace

CouldntExpandBranchHash ShortCausalHash

AmbiguousBranchHash ShortCausalHash (Set ShortCausalHash)

AmbiguousHashForDefinition ShortHash

NoBranchForHash CausalHash

CouldntLoadBranch CausalHash

MissingSignatureForTerm Reference

NoSuchDefinition (HashQualified Name)

ExpectedNameLookup BranchHash

DisjointProjectAndPerspective Path Path

ProjectBranchNameNotFound ProjectName ProjectBranchName

newtype Backend m a

Constructors

Backend

Fields

runBackend :: ReaderT BackendEnv (ExceptT BackendError m) a

data ShallowListEntry v a

Constructors

ShallowTermEntry (TermEntry v a)

ShallowTypeEntry TypeEntry

ShallowBranchEntry NameSegment CausalHash NamespaceStats

ShallowPatchEntry NameSegment

listEntryName :: ShallowListEntry v a -> Text

newtype BackendEnv

Constructors

BackendEnv

Fields

useNamesIndex :: Bool

Whether to use the sqlite name-lookup table to generate Names objects rather than building Names from the root branch.

data TermEntry v a

Constructors

TermEntry

Fields

termEntryReferent :: Referent

termEntryHash :: ShortHash

termEntryName :: Name

termEntryConflicted :: Bool

termEntryType :: Maybe (Type v a)

termEntryTag :: TermTag

data TypeEntry

Constructors

TypeEntry

Fields

typeEntryReference :: Reference

typeEntryHash :: ShortHash

typeEntryName :: Name

typeEntryConflicted :: Bool

typeEntryTag :: TypeTag

data FoundRef

Constructors

FoundTermRef Referent

FoundTypeRef Reference

data IncludeCycles

Determines whether we include full cycles in the results, (e.g. if I search for isEven, will I find isOdd too?)

This was once used for both term and decl components, but now is only used for decl components, because update does The Right Thing for terms (i.e. propagates changes to all dependents, including component-mates, which are de facto dependents).

Ticket of interest: https://github.com/unisonweb/unison/issues/3445

Constructors

IncludeCycles

DontIncludeCycles

data DefinitionResults

Constructors

DefinitionResults

Fields

termResults :: Map Reference (DisplayObject (Type Symbol Ann) (Term Symbol Ann))

typeResults :: Map Reference (DisplayObject () (Decl Symbol Ann))

noResults :: [HashQualified Name]

type SyntaxText = SyntaxText' Reference

Endpoints

fuzzyFind :: Names -> String -> [(Alignment, UnisonName, [FoundRef])]

Utilities

bestNameForTerm :: forall v. Var v => PrettyPrintEnv -> Width -> Referent -> Text

bestNameForType :: forall v. Var v => PrettyPrintEnv -> Width -> Reference -> Text

definitionsByName :: Codebase m Symbol Ann -> NameSearch Transaction -> IncludeCycles -> SearchType -> [HashQualified Name] -> Transaction DefinitionResults

displayType :: Codebase m Symbol Ann -> Reference -> Transaction (DisplayObject () (Decl Symbol Ann))

docsInBranchToHtmlFiles :: Runtime Symbol -> Codebase IO Symbol Ann -> Branch IO -> FilePath -> IO [Error]

expandShortCausalHash :: ShortCausalHash -> Backend Transaction CausalHash

findDocInBranch :: Set NameSegment -> Branch m -> Maybe TermReference

formatSuffixedType :: Var v => PrettyPrintEnvDecl -> Width -> Type v Ann -> SyntaxText

getShallowCausalAtPathFromRootHash :: CausalHash -> Path -> Transaction (CausalBranch Transaction)

Efficiently resolve a root hash and path to a shallow branch's causal.

getTermTag :: (Var v, MonadIO m) => Codebase m v a -> Referent -> Maybe (Type v Ann) -> m TermTag

getTypeTag :: Var v => Codebase m v Ann -> Reference -> Transaction TypeTag

hoistBackend :: (forall x. m x -> n x) -> Backend m a -> Backend n a

hqNameQuery :: Codebase m v Ann -> NameSearch Transaction -> SearchType -> [HashQualified Name] -> Transaction QueryResult

loadReferentType :: Codebase m Symbol Ann -> Referent -> Transaction (Maybe (Type Symbol Ann))

loadSearchResults :: Codebase m Symbol Ann -> [SearchResult] -> Transaction [SearchResult' Symbol Ann]

lsAtPath :: MonadIO m => Codebase m Symbol Ann -> Branch Transaction -> Absolute -> m [ShallowListEntry Symbol Ann]

lsBranch :: MonadIO m => Codebase m Symbol Ann -> Branch n -> m [ShallowListEntry Symbol Ann]

Find all definitions and children reachable from the given Branch,

mungeSyntaxText :: Functor g => g (Element Reference) -> g Element

expectCausalBranchByCausalHash :: CausalHash -> Transaction (CausalBranch Transaction)

resolveRootBranchHashV2 :: ShortCausalHash -> Backend Transaction (CausalBranch Transaction)

namesAtPathFromRootBranchHash :: forall m n v a. MonadIO m => Codebase m v a -> CausalBranch n -> Path -> Backend m (Names, PrettyPrintEnvDecl)

Gets the names and PPED for the branch at the provided path from the root branch for the provided branch hash.

termEntryDisplayName :: TermEntry v a -> Text

termEntryHQName :: TermEntry v a -> HashQualified Name

termEntryToNamedTerm :: Var v => PrettyPrintEnv -> Maybe Width -> TermEntry v a -> NamedTerm

termEntryLabeledDependencies :: Ord v => TermEntry v a -> Set LabeledDependency

termListEntry :: MonadIO m => Codebase m Symbol Ann -> ExactName Name Referent -> m (TermEntry Symbol Ann)

termReferentsByShortHash :: Codebase m v a -> ShortHash -> Transaction (Set Referent)

Look up terms in the codebase by short hash, and include builtins.

typeDeclHeader :: forall v m. Var v => Codebase m v Ann -> PrettyPrintEnv -> Reference -> Transaction (DisplayObject SyntaxText SyntaxText)

typeEntryDisplayName :: TypeEntry -> Text

typeEntryHQName :: TypeEntry -> HashQualified Name

typeEntryToNamedType :: TypeEntry -> NamedType

typeEntryLabeledDependencies :: TypeEntry -> Set LabeledDependency

typeListEntry :: Var v => Codebase m v Ann -> ExactName Name Reference -> Transaction TypeEntry

typeReferencesByShortHash :: ShortHash -> Transaction (Set Reference)

Look up types in the codebase by short hash, and include builtins.

typeToSyntaxHeader :: Width -> HashQualified Name -> DisplayObject () (Decl Symbol Ann) -> DisplayObject SyntaxText SyntaxText

Renders a type to its decl header, e.g.

Effect:

unique ability Stream s

Data:

structural type Maybe a

renderDocRefs :: Traversable t => PrettyPrintEnvDecl -> Width -> Codebase IO Symbol Ann -> Runtime Symbol -> t TermReference -> IO (t (HashQualifiedName, UnisonHash, Doc, [Error]))

Evaluate and render the given docs

docsForDefinitionName :: Codebase IO Symbol Ann -> NameSearch Transaction -> SearchType -> Name -> Transaction [TermReference]

Fetch the docs associated with the given name. Returns all references with a Doc type which are at the name provided, or at 'name.doc'.

normaliseRootCausalHash :: Either ShortCausalHash CausalHash -> Backend Transaction (CausalBranch Transaction)

Unused, could remove?

resolveRootBranchHash :: MonadIO m => ShortCausalHash -> Codebase m v a -> Backend m (Branch m)

isTestResultList :: forall v a. (Var v, Monoid a) => Maybe (Type v a) -> Bool

fixupNamesRelative :: Absolute -> Names -> Names

Re-exported for Share Server

termsToSyntax :: Var v => Ord a => Suffixify -> Width -> PrettyPrintEnvDecl -> [(TermReference, DisplayObject (Type v a) (Term v a))] -> [(TermReference, DisplayObject SyntaxText SyntaxText)]

Converts Type Display Objects into Syntax Text.

termsToSyntaxOf :: Var v => Ord a => Suffixify -> Width -> PrettyPrintEnvDecl -> Traversal s t (TermReference, DisplayObject (Type v a) (Term v a)) (TermReference, DisplayObject SyntaxText SyntaxText) -> s -> t

Version of termsToSyntax which works over arbitrary traversals.

E.g. @@ termsToSyntaxOf suff width pped traversed [(ref, dispObj)]

or

termsToSyntaxOf suff width pped id (ref, dispObj)

or

termsToSyntaxOf suff width pped Map.asList_ (Map.singleton ref dispObj) @@ e.g. traversed

typesToSyntax :: Var v => Ord a => Suffixify -> Width -> PrettyPrintEnvDecl -> [(TypeReference, DisplayObject () (Decl v a))] -> [(TypeReference, DisplayObject SyntaxText SyntaxText)]

Converts Type Display Objects into Syntax Text.

typesToSyntaxOf :: Var v => Ord a => Suffixify -> Width -> PrettyPrintEnvDecl -> Traversal s t (TypeReference, DisplayObject () (Decl v a)) (TypeReference, DisplayObject SyntaxText SyntaxText) -> s -> t

Version of typesToSyntax which works over arbitrary traversals.

E.g. @@ typesToSyntaxOf suff width pped traversed [(ref, dispObj)]

or

typesToSyntaxOf suff width pped id (ref, dispObj)

or

typesToSyntaxOf suff width pped Map.asList_ (Map.singleton ref dispObj) @@

definitionResultsDependencies :: DefinitionResults -> Set LabeledDependency

Finds ALL direct references contained within a DefinitionResults so we can build a pretty printer for them.

evalDocRef :: Runtime Symbol -> Codebase IO Symbol Ann -> TermReference -> IO (EvaluatedDoc Symbol, [Error])

Evaluate the doc at the given reference and return its evaluated-but-not-rendered form.

mkTermDefinition :: Codebase IO Symbol Ann -> PrettyPrintEnvDecl -> Width -> Reference -> [(HashQualifiedName, UnisonHash, Doc)] -> DisplayObject (AnnotatedText (Element Reference)) (AnnotatedText (Element Reference)) -> Backend IO TermDefinition

mkTypeDefinition :: MonadIO m => Codebase IO Symbol Ann -> PrettyPrintEnvDecl -> Width -> Reference -> [(HashQualifiedName, UnisonHash, Doc)] -> DisplayObject (AnnotatedText (Element Reference)) (AnnotatedText (Element Reference)) -> m TypeDefinition

displayTerm :: Codebase m Symbol Ann -> Reference -> Transaction (DisplayObject (Type Symbol Ann) (Term Symbol Ann))

formatTypeName :: PrettyPrintEnv -> Reference -> SyntaxText