Unison.Server.Doc

Documentation

type Nat = Word64

type SSyntaxText = SyntaxText' Reference

type Doc = DocG RenderedSpecialForm

A doc rendered down to SyntaxText.

type EvaluatedDoc v = DocG (EvaluatedSpecialForm v)

A doc which has been evaluated and includes all information necessary to be rendered.

type SrcRefs = Ref (UnisonHash, DisplayObject SyntaxText Src)

data DocG specialForm

A doc parameterized by its special forms.

Constructors

Word Text

Code (DocG specialForm)

CodeBlock Text (DocG specialForm)

Bold (DocG specialForm)

Italic (DocG specialForm)

Strikethrough (DocG specialForm)

Style Text (DocG specialForm)

Anchor Text (DocG specialForm)

Blockquote (DocG specialForm)

Blankline

Linebreak

SectionBreak

Tooltip (DocG specialForm) (DocG specialForm)

Aside (DocG specialForm)

Callout (Maybe (DocG specialForm)) (DocG specialForm)

Table [[DocG specialForm]]

Folded Bool (DocG specialForm) (DocG specialForm)

Paragraph [DocG specialForm]

BulletedList [DocG specialForm]

NumberedList Nat [DocG specialForm]

Section (DocG specialForm) [DocG specialForm]

NamedLink (DocG specialForm) (DocG specialForm)

Image (DocG specialForm) (DocG specialForm) (Maybe (DocG specialForm))

Special specialForm

Join [DocG specialForm]

UntitledSection [DocG specialForm]

Column [DocG specialForm]

Group (DocG specialForm)

type UnisonHash = Text

data Ref a

Constructors

Term a

Type a

data MediaSource

Constructors

MediaSource

Fields

mediaSourceUrl :: Text

mediaSourceMimeType :: Maybe Text

data RenderedSpecialForm

Constructors

Source [SrcRefs]

FoldedSource [SrcRefs]

Example SyntaxText

ExampleBlock SyntaxText

Link SyntaxText

Signature [SyntaxText]

SignatureInline SyntaxText

Eval SyntaxText SyntaxText

EvalInline SyntaxText SyntaxText

Embed SyntaxText

EmbedInline SyntaxText

Video [MediaSource] (Map Text Text)

FrontMatter (Map Text [Text])

LaTeXInline Text

Svg Text

RenderError (RenderError SyntaxText)

data EvaluatedSpecialForm v

Constructors

ESource [EvaluatedSrc v]

EFoldedSource [EvaluatedSrc v]

EExample (Term v ())

EExampleBlock (Term v ())

ELink (Either (Term v ()) LabeledDependency)

ESignature [(Referent, Type v ())]

ESignatureInline (Referent, Type v ())

EEval (Term v ()) (Maybe (Term v ()))

EEvalInline (Term v ()) (Maybe (Term v ()))

EEmbed (Term v ())

EEmbedInline (Term v ())

EVideo [MediaSource] (Map Text Text)

EFrontMatter (Map Text [Text])

ELaTeXInline Text

ESvg Text

ERenderError (RenderError (Term v ()))

data Src

Constructors

Src SyntaxText SyntaxText

evalAndRenderDoc :: forall v m. (Var v, Monad m) => PrettyPrintEnvDecl -> (Reference -> m (Maybe (Term v ()))) -> (Referent -> m (Maybe (Type v ()))) -> (Term v () -> m (Maybe (Term v ()))) -> (Reference -> m (Maybe (Decl v ()))) -> Term v () -> m Doc

Evaluate the doc, then render it.

renderDoc :: forall v. Var v => PrettyPrintEnvDecl -> EvaluatedDoc v -> Doc

Renders the given doc, which must have been evaluated using evalDoc

evalDoc :: forall v m. (Var v, Monad m) => (Reference -> m (Maybe (Term v ()))) -> (Referent -> m (Maybe (Type v ()))) -> (Term v () -> m (Maybe (Term v ()))) -> (Reference -> m (Maybe (Decl v ()))) -> Term v () -> m (EvaluatedDoc v)

Evaluates the given doc, expanding transclusions, expressions, etc.

data RenderError trm

Constructors

InvalidTerm trm

data EvaluatedSrc v

Constructors

EvaluatedSrcDecl (EvaluatedDecl v)

EvaluatedSrcTerm (EvaluatedTerm v)

data EvaluatedDecl v

Constructors

MissingDecl Reference

BuiltinDecl Reference

FoundDecl Reference (Decl v ())

data EvaluatedTerm v

Constructors

MissingTerm Reference

BuiltinTypeSig Reference (Type v ())

MissingBuiltinTypeSig Reference

FoundTerm Reference (Type v ()) (Term v ())

dependencies :: Ord v => EvaluatedDoc v -> Set LabeledDependency

dependenciesSpecial :: forall v. Ord v => EvaluatedSpecialForm v -> Set LabeledDependency

Determines all dependencies of a special form