Unison.Server.Doc.AsHtml

Description

Render Unison.Server.Doc and embedded source to Html

Documentation

data NamedLinkHref

Constructors

Href Text

DocLinkHref Name

ReferenceHref Text

InvalidHref

data EmbeddedSource

Constructors

EmbeddedSource SyntaxText SyntaxText

Builtin SyntaxText

embeddedSource :: Ref (UnisonHash, DisplayObject SyntaxText Src) -> Maybe EmbeddedSource

inlineCode :: [Text] -> Html () -> Html ()

codeBlock :: [Attribute] -> Html () -> Html ()

normalizeHref :: Map Referent Name -> Doc -> NamedLinkHref

data IsFolded

Constructors

IsFolded Bool [Html ()] [Html ()]

Disabled (Html ())

foldedToHtml :: [Attribute] -> IsFolded -> Html ()

foldedToHtmlSource :: Bool -> EmbeddedSource -> Html ()

mergeWords :: Text -> [Doc] -> [Doc]

Merge adjacent Word elements in a list to 1 element with a string of words separated by space— useful for rendering to the dom without creating dom elements for each and every word in the doc, but instead rely on textNodes

toText :: Text -> Doc -> Text

Merge down Doc to Text by merging Paragraphs and Words. Used for things like extract an src of an image. I.e something that has to be a Text and not a Doc

data SideContent

Constructors

FrontMatterContent (Map Text [Text])

TooltipContent (Html ())

newtype FrontMatterData

Constructors

FrontMatterData (Map Text [Text])

toHtml :: Map Referent Name -> Doc -> (FrontMatterData, Html ())

h :: Nat -> Text -> Html () -> Html ()

Unison Doc allows endlessly deep section nesting with titles, but HTML only supports to h1-h6, so we clamp the sectionLevel when converting

badge :: Html () -> Html ()

textToClass :: Text -> Text