Unison.Server.Local

Description

Helpers which are specific to the local share server.

Documentation

relocateToNameRoot :: Path -> HashQualified Name -> Branch Transaction -> Transaction (Either BackendError (Path, HashQualified Name))

Given an arbitrary query and perspective, find the name root the query belongs in, then return that root and the query relocated to that root.

A name root is either a project root or a dependency root. E.g. .myproject.some.namespace -> .myproject or .myproject.lib.base.List -> .myproject.lib.base