Unison.Server.Local.Definitions

Documentation

prettyDefinitionsForHQName

Arguments

:: Path
The path representing the user's current perspective. Searches will be limited to definitions within this path, and names will be relative to this path.

-> CausalBranch Transaction
The root branch to use

-> Maybe Width

-> Suffixify
Whether to suffixify bindings in the rendered syntax

-> Runtime Symbol
Runtime used to evaluate docs. This should be sandboxed if run on the server.

-> Codebase IO Symbol Ann

-> HashQualified Name
The name, hash, or both, of the definition to display.

-> Backend IO DefinitionDisplayResults

Renders a definition for the given name or hash alongside its documentation.

termDefinitionByName :: Codebase IO Symbol Ann -> PrettyPrintEnvDecl -> NameSearch Transaction -> Width -> Runtime Symbol -> Name -> Backend IO (Maybe TermDefinition)

typeDefinitionByName :: Codebase IO Symbol Ann -> PrettyPrintEnvDecl -> NameSearch Transaction -> Width -> Runtime Symbol -> Name -> Backend IO (Maybe TypeDefinition)