Unison.Server.Local.Endpoints.Projects.Queries

Documentation

listProjects :: Maybe Text -> Transaction [ProjectListing]

Load all project listings, optionally requiring an infix match with a query.

listProjectBranches :: ProjectId -> Maybe Text -> Transaction [ProjectBranchListing]

Load all project listings, optionally requiring an infix match with a query.