Unison.Server.NameSearch

Documentation

data Search m r

A Search r is a small bag of functions that is used to power a search for rs.

There are currently two implementations of this interface:

FromNames which builds a name search from a Names object

'NameSearch.Sqlite which builds a name search that uses a sqlite name lookup index.

You can use the individual methods of a name search or can use applySearch.

Constructors

Search

Fields

lookupNames :: r -> m (Set (HashQualified Name))

lookupRelativeHQRefs' :: SearchType -> HashQualified Name -> m (Set r)

makeResult :: HashQualified Name -> r -> Set (HashQualified Name) -> m SearchResult

matchesNamedRef :: Name -> r -> HashQualified Name -> Bool

data NameSearch m

Constructors

NameSearch

Fields

typeSearch :: Search m Reference

termSearch :: Search m Referent

hoistSearch :: (forall x. m x -> n x) -> Search m r -> Search n r

hoistNameSearch :: (forall x. m x -> n x) -> NameSearch m -> NameSearch n

applySearch :: (Show r, Monad m) => Search m r -> SearchType -> HashQualified Name -> m [SearchResult]

Interpret a Search as a function from name to search results.

data SearchType

Whether to search for exact matches or to find definitions by a suffix of their name.

Constructors

IncludeSuffixes

ExactName