Unison.Server.NameSearch.FromNames

Documentation

makeTypeSearch :: Applicative m => Int -> Names -> Search m Reference

Make a type search, given a short hash length and names to search in.

makeTermSearch :: Applicative m => Int -> Names -> Search m Referent

Make a term search, given a short hash length and names to search in.

makeNameSearch :: Applicative m => Int -> Names -> NameSearch m