Unison.Server.NameSearch.Sqlite

Documentation

resolveShortHash :: Codebase m v a -> ShortHash -> Transaction (Set LabeledDependency)

Resolves a shorthash into any possible matches.

typeReferencesByShortHash :: ShortHash -> Transaction (Set Reference)

Look up types in the codebase by short hash, and include builtins.

termReferentsByShortHash :: Codebase m v a -> ShortHash -> Transaction (Set Referent)

Look up terms in the codebase by short hash, and include builtins.

data NameSearch m

Constructors

NameSearch

Fields

typeSearch :: Search m Reference

termSearch :: Search m Referent

nameSearchForPerspective :: Codebase m v a -> NamesPerspective -> NameSearch Transaction