Unison.Server.SearchResult

Documentation

data SearchResult

Constructors

Tp TypeResult

Tm TermResult

data TermResult

Constructors

TermResult

Fields

termName :: HashQualified Name

referent :: Referent

termAliases :: Set (HashQualified Name)

data TypeResult

Constructors

TypeResult

Fields

typeName :: HashQualified Name

reference :: Reference

typeAliases :: Set (HashQualified Name)

pattern Tm' :: HashQualified Name -> Referent -> Set (HashQualified Name) -> SearchResult

pattern Tp' :: HashQualified Name -> Reference -> Set (HashQualified Name) -> SearchResult

termResult :: HashQualified Name -> Referent -> Set (HashQualified Name) -> SearchResult

Construct a term search result from a primary name, referent, and set of aliases.

termSearchResult :: Names -> Name -> Referent -> SearchResult

typeResult :: HashQualified Name -> Reference -> Set (HashQualified Name) -> SearchResult

Construct a type search result from a primary name, reference, and set of aliases.

typeSearchResult :: Names -> Name -> Reference -> SearchResult

name :: SearchResult -> HashQualified Name

aliases :: SearchResult -> Set (HashQualified Name)

toReferent :: SearchResult -> Referent

TypeResults yield a Ref

truncateAliases :: Int -> SearchResult -> SearchResult

fromNames :: Names -> [SearchResult]

You may want to sort this list differently afterward.

_fromNames :: Names -> [SearchResult]

compareByName :: SearchResult -> SearchResult -> Ordering

Sort a list of search results by name. If names are equal, fall back to comparing by reference (putting types before terms).