Unison.Server.Share.Definitions

Description

This module contains implementations of Backend methods which are specialized for Share. We should likely move them to the Share repository eventually, but for now it's much easier to ensure they're resilient to refactors and changes in the Backend API if they live here.

Perhaps we'll move them when the backing implementation switches to postgres.

Documentation

definitionForHQName

Arguments

:: Path
The path representing the user's current namesRoot. Searches will be limited to definitions within this path, and names will be relative to this path.

-> CausalHash
The root branch to use

-> Maybe Width

-> Suffixify
Whether to suffixify bindings in the rendered syntax

-> Runtime Symbol
Runtime used to evaluate docs. This should be sandboxed if run on the server.

-> Codebase IO Symbol Ann

-> HashQualified Name
The name, hash, or both, of the definition to display.

-> Backend IO DefinitionDisplayResults

Renders a definition for the given name or hash alongside its documentation.