Unison.Server.Share.FuzzyFind

Documentation

type FuzzyFindAPI = "find" :> (QueryParam "rootBranch" ShortCausalHash :> (QueryParam "relativeTo" Path :> (QueryParam "limit" Int :> (QueryParam "renderWidth" Width :> (QueryParam "query" String :> APIGet [(Alignment, FoundResult)])))))

data FoundTerm

Constructors

FoundTerm

Fields

bestFoundTermName :: HashQualifiedName

namedTerm :: NamedTerm

data FoundType

Constructors

FoundType

Fields

bestFoundTypeName :: HashQualifiedName

typeDef :: DisplayObject SyntaxText SyntaxText

namedType :: NamedType

data FoundResult

Constructors

FoundTermResult FoundTerm

FoundTypeResult FoundType

serveFuzzyFind

Arguments

:: Bool
Whether the root is a scratch root

-> Bool
Whether to search in dependencies

-> Codebase IO Symbol Ann

-> CausalHash

-> Path

-> Maybe Int

-> Maybe Width

-> Text

-> Backend IO [(Alignment, FoundResult)]

scoreMatch :: NamedRef r -> Int

data Alignment

Constructors

Alignment

Fields

score :: Int

result :: MatchResult

data MatchResult

Constructors

MatchResult

Fields

segments :: [MatchSegment]

data MatchSegment

Constructors

Gap Text

Match Text

computeMatchSegments :: [Text] -> NamedRef r -> [MatchSegment]

prepareQuery :: String -> [Text]

Splits a query into segments, where each segment must appear in order in any matching names.

>>> prepareQuery "foo bar baz"
["foo","bar","baz"]

Split camel-case style words into segments. >>> prepareQuery "fMap" ["f",Map]

Collapse multiple spaces >>> prepareQuery "foo barBaz boom" ["foo","bar",Baz,"boom"]

Split namespaces into segments with a required dot in between. >>> prepareQuery "List.map" [List,".","map"]

Shouldn't get multiple splits for capitalized letters >>> prepareQuery List.Map [List,".",Map]

Orphan instances

ToParam (QueryParam "limit" Int)

Instance details

Methods

toParam :: Proxy (QueryParam "limit" Int) -> DocQueryParam

ToParam (QueryParam "query" String)

Instance details

Methods

toParam :: Proxy (QueryParam "query" String) -> DocQueryParam