Unison.Server.Share.RenderDoc

Description

Helper for rendering docs within a given namespace

Documentation

findAndRenderDoc :: Set NameSegment -> Runtime Symbol -> Codebase IO Symbol Ann -> Path -> CausalHash -> Maybe Width -> Backend IO (Maybe Doc)

Find, eval, and render the first doc we find with any of the provided names within the given namespace If no doc is found, return Nothing

Requires Name Lookups, currently only usable on Share.