Unison.Server.Syntax

Description

Duplicate of the Unison.Util.SyntaxText module, but we expect these to evolve separately. This is the version which is outward facing to the server frontend.

Documentation

type SyntaxText = AnnotatedText Element

type SyntaxSegment = Segment Element

convertElement :: Element Reference -> Element

type UnisonHash = Text

type HashQualifiedName = Text

data Element

The elements of the Unison grammar, for syntax highlighting purposes

Constructors

NumericLiteral

TextLiteral

BytesLiteral

CharLiteral

BooleanLiteral

Blank

Var

TypeReference UnisonHash

DataConstructorReference UnisonHash

AbilityConstructorReference UnisonHash

TermReference UnisonHash

Op SeqOp

AbilityBraces
Constructor Are these even used? | Request

ControlKeyword

TypeOperator

BindingEquals

TypeAscriptionColon

DataTypeKeyword

DataTypeParams

Unit

DataTypeModifier

UseKeyword

UsePrefix

UseSuffix

HashQualifier HashQualifiedName

DelayForceChar

DelimiterChar

Parenthesis

LinkKeyword

DocDelimiter

DocKeyword

syntax :: Element -> SyntaxText -> SyntaxText

firstReference :: SyntaxText -> Maybe UnisonHash

reference :: SyntaxSegment -> Maybe UnisonHash

toPlain :: SyntaxText -> String

Convert a SyntaxText to a Value, ignoring syntax markup

toPlainText :: SyntaxText -> Text

toHtml :: SyntaxText -> Html ()

nameToHtml :: Name -> Html ()

segmentToHtml :: SyntaxSegment -> Html ()

elementToClassName :: Element -> Text

Orphan instances

FromJSON SeqOp

Instance details

Methods

parseJSON :: Value -> Parser SeqOp

parseJSONList :: Value -> Parser [SeqOp]

ToJSON SeqOp

Instance details

Methods

toJSON :: SeqOp -> Value

toEncoding :: SeqOp -> Encoding

toJSONList :: [SeqOp] -> Value

toEncodingList :: [SeqOp] -> Encoding

ToSchema SeqOp

Instance details

Methods

declareNamedSchema :: Proxy SeqOp -> Declare (Definitions Schema) NamedSchema

FromJSON a => FromJSON (Segment a)

Instance details

Methods

parseJSON :: Value -> Parser (Segment a)

parseJSONList :: Value -> Parser [Segment a]

ToJSON a => ToJSON (Segment a)

Instance details

Methods

toJSON :: Segment a -> Value

toEncoding :: Segment a -> Encoding

toJSONList :: [Segment a] -> Value

toEncodingList :: [Segment a] -> Encoding

ToSchema r => ToSchema (Seq r)

Instance details

Methods

declareNamedSchema :: Proxy (Seq r) -> Declare (Definitions Schema) NamedSchema

ToSchema a => ToSchema (Segment a)

Instance details

Methods

declareNamedSchema :: Proxy (Segment a) -> Declare (Definitions Schema) NamedSchema