Unison.Server.Types

Documentation

type APIHeaders x = Headers '[Header "Cache-Control" String] x

type APIGet c = Get '[JSON] (APIHeaders c)

type HashQualifiedName = Text

type NamespaceFQN = Text

type Size = Int

type UnisonName = Text

type UnisonHash = Text

data NamespaceDetails

Constructors

NamespaceDetails

Fields

fqn :: Path

hash :: UnisonHash

readme :: Maybe Doc

data ExactName name ref

A hash qualified name, unlike HashQualified, the hash is required

Constructors

ExactName

Fields

name :: name

ref :: ref

exactToHQ :: ExactName name ShortHash -> HashQualified name

exactToHQ' :: ExactName name ShortHash -> HashQualified name

data TermDefinitionDiff

Constructors

TermDefinitionDiff

Fields

left :: TermDefinition

right :: TermDefinition

diff :: DisplayObjectDiff

data TypeDefinitionDiff

Constructors

TypeDefinitionDiff

Fields

left :: TypeDefinition

right :: TypeDefinition

diff :: DisplayObjectDiff

newtype Suffixify

Constructors

Suffixify

Fields

suffixified :: Bool

data TermDefinition

Constructors

TermDefinition

Fields

termNames :: [HashQualifiedName]

bestTermName :: HashQualifiedName

defnTermTag :: TermTag

termDefinition :: DisplayObject SyntaxText SyntaxText

signature :: SyntaxText

termDocs :: [(HashQualifiedName, UnisonHash, Doc)]

data TypeDefinition

Constructors

TypeDefinition

Fields

typeNames :: [HashQualifiedName]

bestTypeName :: HashQualifiedName

defnTypeTag :: TypeTag

typeDefinition :: DisplayObject SyntaxText SyntaxText

typeDocs :: [(HashQualifiedName, UnisonHash, Doc)]

data DefinitionDisplayResults

Constructors

DefinitionDisplayResults

Fields

termDefinitions :: Map UnisonHash TermDefinition

typeDefinitions :: Map UnisonHash TypeDefinition

missingDefinitions :: [HashQualifiedName]

data TermTag

Constructors

Doc

Test

Plain

Constructor TypeTag

data TypeTag

Constructors

Ability

Data

data SemanticSyntaxDiff a

A type for semantic diffing of definitions. Includes special-cases for when the name in a definition has changed but the hash hasn't (rename/alias), and when the hash has changed but the name hasn't (update propagation).

Constructors

OnlyThisSide (NonEmpty (Segment a))

Both (NonEmpty (Segment a))

SegmentChange (Text, Text) (Maybe a)

AnnotationChange Text (Maybe a, Maybe a)

data Changed a

Constructors

Changed a

Unchanged a

Spacer

data LinewiseDiff a

Constructors

LinewiseDiff

Fields

lhsLines :: [Changed [a]]

rhsLines :: [Changed [a]]

data Paired a

Constructors

OneSided a

Paired a a

swapPair :: Paired a -> Paired a

data DisplayObjectDiff

A diff of the syntax of a term or type

It doesn't make sense to diff builtins with ABTs, so in that case we just provide the undiffed syntax.

Constructors

DisplayObjectDiff (DisplayObject (LinewiseDiff (SemanticSyntaxDiff Element)) (LinewiseDiff (SemanticSyntaxDiff Element)))

MismatchedDisplayObjects (DisplayObject SyntaxText SyntaxText) (DisplayObject SyntaxText SyntaxText)

data UnisonRef

Constructors

TypeRef UnisonHash

TermRef UnisonHash

unisonRefToText :: UnisonRef -> Text

data NamedTerm

Constructors

NamedTerm

Fields

termName :: HashQualified Name

termHash :: ShortHash

termType :: Maybe SyntaxText

termTag :: TermTag

data NamedType

Constructors

NamedType

Fields

typeName :: HashQualified Name

typeHash :: ShortHash

typeTag :: TypeTag

munge :: Text -> ByteString

mungeShow :: Show s => s -> ByteString

mungeString :: String -> ByteString

defaultWidth :: Width

discard :: Applicative m => a -> m ()

mayDefaultWidth :: Maybe Width -> Width

setCacheControl :: v -> APIHeaders v

branchToUnisonHash :: forall (m :: Type -> Type). Branch m -> UnisonHash

v2CausalBranchToUnisonHash :: forall (m :: Type -> Type). CausalBranch m -> UnisonHash

newtype ProjectBranchNameParam

Constructors

ProjectBranchNameParam

Fields

unProjectBranchNameParam :: ProjectAndBranch ProjectName ProjectBranchName

data TermDiffResponse

Constructors

TermDiffResponse

Fields

project :: ProjectName

oldBranch :: ProjectBranchName

newBranch :: ProjectBranchName

oldTerm :: TermDefinition

newTerm :: TermDefinition

diff :: DisplayObjectDiff

data TypeDiffResponse

Constructors

TypeDiffResponse

Fields

project :: ProjectName

oldBranch :: ProjectBranchName

newBranch :: ProjectBranchName

oldType :: TypeDefinition

newType :: TypeDefinition

diff :: DisplayObjectDiff

type RequiredQueryParam = QueryParam' '[Required, Strict]

Servant utility for a query param that's required, providing a useful error message if it's missing.

data DefinitionNameSearchResult

Constructors

DefinitionNameSearchResult

Fields

token :: Name

tag :: TermOrTypeTag

newtype DefinitionSearchResults

Constructors

DefinitionSearchResults

Fields

results :: [DefinitionSearchResult]

data DefinitionSearchResult

Constructors

DefinitionSearchResult

Fields

fqn :: Name

summary :: TermOrTypeSummary

project :: ProjectName

branchRef :: ProjectBranchName

data TermSummary

Constructors

TermSummary

Fields

displayName :: HashQualified Name

hash :: ShortHash

summary :: DisplayObject SyntaxText SyntaxText

tag :: TermTag

data TypeSummary

Constructors

TypeSummary

Fields

displayName :: HashQualified Name

hash :: ShortHash

summary :: DisplayObject SyntaxText SyntaxText

tag :: TypeTag

data TermOrTypeSummary

Constructors

ToTTermSummary TermSummary

ToTTypeSummary TypeSummary

data TermOrTypeTag

Constructors

ToTTermTag TermTag

ToTTypeTag TypeTag

newtype BranchRef

Constructors

BranchRef

Fields

unBranchRef :: Text