Unison.Sync.Common

Description

Combinators or utilities shared by sync server AND client

Documentation

expectEntity :: Hash32 -> Transaction (Entity Text Hash32 Hash32)

Read an entity out of the database that we know is in main storage.

Type conversions

causalHashToHash32 :: CausalHash -> Hash32

hash32ToCausalHash :: Hash32 -> CausalHash

entityToTempEntity :: forall hash. (hash -> Hash32) -> Entity Text Hash32 hash -> TempEntity

Convert an entity that came over the wire from Unison Share into an equivalent type that we can store in the temp_entity table.

tempEntityToEntity :: TempEntity -> Entity Text Hash32 Hash32