Unison.Sync.Types

Misc. types

newtype Base64Bytes

A newtype for JSON encoding binary data.

Constructors

Base64Bytes ByteString

newtype RepoName

Constructors

RepoName

Fields

unRepoName :: Text

data Path

Constructors

Path

Fields

pathSegments :: NonEmpty Text

pathRepoName :: Path -> RepoName

pathCodebasePath :: Path -> [Text]

Hash types

newtype HashJWT

Constructors

HashJWT

Fields

unHashJWT :: Text

hashJWTHash :: HashJWT -> Hash32

Grab the hash out of a hash JWT.

This decodes the whole JWT, then throws away the claims; use it if you really only need the hash!

data HashJWTClaims

Constructors

HashJWTClaims

Fields

hash :: Hash32

userId :: Maybe Text

data DecodedHashJWT

A decoded hash JWT that retains the original encoded JWT.

Constructors

DecodedHashJWT

Fields

claims :: HashJWTClaims

hashJWT :: HashJWT

decodeHashJWT :: HashJWT -> DecodedHashJWT

Decode a hash JWT.

decodeHashJWTClaims :: HashJWT -> HashJWTClaims

Decode the claims out of a hash JWT.

decodedHashJWTHash :: DecodedHashJWT -> Hash32

Grab the hash out of a decoded hash JWT.

Entity types

data Entity text noSyncHash hash

Constructors

TC (TermComponent text hash)

DC (DeclComponent text hash)

P (Patch text noSyncHash hash)

PD (PatchDiff text noSyncHash hash)

N (Namespace text hash)

ND (NamespaceDiff text hash)

C (Causal hash)

data TermComponent text hash

Constructors

TermComponent [(LocalIds text hash, ByteString)]

data DeclComponent text hash

Constructors

DeclComponent [(LocalIds text hash, ByteString)]

data Patch text oldHash newHash

Constructors

Patch

Fields

textLookup :: [text]

oldHashLookup :: [oldHash]

newHashLookup :: [newHash]

bytes :: ByteString

data PatchDiff text oldHash hash

Constructors

PatchDiff

Fields

parent :: hash

textLookup :: [text]

oldHashLookup :: [oldHash]

newHashLookup :: [hash]

bytes :: ByteString

data Namespace text hash

Constructors

Namespace

Fields

textLookup :: [text]

defnLookup :: [hash]

patchLookup :: [hash]

childLookup :: [(hash, hash)]

bytes :: ByteString

data NamespaceDiff text hash

Constructors

NamespaceDiff

Fields

parent :: hash

textLookup :: [text]

defnLookup :: [hash]

patchLookup :: [hash]

childLookup :: [(hash, hash)]

bytes :: ByteString

data Causal hash

Constructors

Causal

Fields

namespaceHash :: hash

parents :: Set hash

data LocalIds text hash

Constructors

LocalIds

Fields

texts :: [text]

hashes :: [hash]

entityDependencies :: Ord hash => Entity text noSyncHash hash -> Set hash

Get the direct dependencies of an entity (which are actually sync'd).

FIXME use generic-lens here? (typed @hash)

data EntityType

Constructors

TermComponentType

DeclComponentType

PatchType

PatchDiffType

NamespaceType

NamespaceDiffType

CausalType

Entity Traversals

entityHashes_ :: (Applicative m, Ord hash') => (hash -> m hash') -> Entity text noSyncHash hash -> m (Entity text noSyncHash hash')

patchHashes_ :: Applicative m => (hash -> m hash') -> Patch text noSyncHash hash -> m (Patch text noSyncHash hash')

patchDiffHashes_ :: Applicative m => (hash -> m hash') -> PatchDiff text noSyncHash hash -> m (PatchDiff text noSyncHash hash')

namespaceDiffHashes_ :: Applicative m => (hash -> m hash') -> NamespaceDiff text hash -> m (NamespaceDiff text hash')

causalHashes_ :: (Applicative m, Ord hash') => (hash -> m hash') -> Causal hash -> m (Causal hash')

Request/response types

Get causal hash by path

newtype GetCausalHashByPathRequest

Constructors

GetCausalHashByPathRequest

Fields

path :: Path

data GetCausalHashByPathResponse

Constructors

GetCausalHashByPathSuccess (Maybe HashJWT)

GetCausalHashByPathNoReadPermission Path

Download entities

data DownloadEntitiesRequest

Constructors

DownloadEntitiesRequest

Fields

repoName :: RepoName

hashes :: NESet HashJWT

data DownloadEntitiesResponse

Constructors

DownloadEntitiesSuccess (NEMap Hash32 (Entity Text Hash32 HashJWT))

DownloadEntitiesNoReadPermission RepoName

Upload entities

data UploadEntitiesRequest

Constructors

UploadEntitiesRequest

Fields

repoName :: RepoName

entities :: NEMap Hash32 (Entity Text Hash32 Hash32)

data UploadEntitiesResponse

Constructors

UploadEntitiesSuccess

UploadEntitiesNeedDependencies (NeedDependencies Hash32)

UploadEntitiesNoWritePermission RepoName

UploadEntitiesHashMismatchForEntity HashMismatchForEntity

Fast-forward path

data FastForwardPathRequest

A non-empty list of causal hashes, latest first, that show the lineage from wherever the client wants to fast-forward to back to wherever the (client believes the) server is (including the server head, in a separate field).

For example, if the client wants to update

A -> B -> C

to

A -> B -> C -> D -> E -> F

then it would send hashes

expectedHash = C
hashes = [D, E, F]

Note that if the client wants to begin a history at a new path on the server, it would use the "update path" endpoint instead.

Constructors

FastForwardPathRequest

Fields

expectedHash :: Hash32

The causal that the client believes exists at $sel:path:GetCausalHashByPathRequest

hashes :: NonEmpty Hash32

The sequence of causals to fast-forward with, starting from the oldest new causal to the newest new causal

path :: Path

The path to fast-forward

data FastForwardPathResponse

Constructors

FastForwardPathSuccess

FastForwardPathMissingDependencies (NeedDependencies Hash32)

FastForwardPathNoWritePermission Path

FastForwardPathNotFastForward HashJWT
This wasn't a fast-forward. Here's a JWT to download the causal head, if you want it.

FastForwardPathNoHistory
There was no history at this path; the client should use the "update path" endpoint instead.

FastForwardPathInvalidParentage InvalidParentage
This wasn't a fast-forward. You said the first hash was a parent of the second hash, but I disagree.

Update path

data UpdatePathRequest

Constructors

UpdatePathRequest

Fields

path :: Path

expectedHash :: Maybe Hash32

newHash :: Hash32

data UpdatePathResponse

Constructors

UpdatePathSuccess

UpdatePathHashMismatch HashMismatch

UpdatePathMissingDependencies (NeedDependencies Hash32)

UpdatePathNoWritePermission Path

data HashMismatch

Constructors

HashMismatch

Fields

path :: Path

expectedHash :: Maybe Hash32

actualHash :: Maybe Hash32

Common/shared error types

data HashMismatchForEntity

Constructors

HashMismatchForEntity

Fields

supplied :: Hash32

computed :: Hash32

data InvalidParentage

Constructors

InvalidParentage

Fields

parent :: Hash32

child :: Hash32

data NeedDependencies hash

Constructors

NeedDependencies

Fields

missingDependencies :: NESet hash