Unison.SyncV2.Types

Documentation

data DownloadEntitiesRequest

Constructors

DownloadEntitiesRequest

Fields

causalHash :: HashJWT

branchRef :: BranchRef

knownHashes :: Set Hash32

data DownloadEntitiesChunk

A chunk of the download entities response stream.

Constructors

InitialC StreamInitInfo

EntityC EntityChunk

ErrorC ErrorChunk

data EntityChunk

Constructors

EntityChunk

Fields

hash :: Hash32

entityCBOR :: CBORBytes TempEntity

data ErrorChunk

Constructors

ErrorChunk

Fields

err :: DownloadEntitiesError

data StreamInitInfo

Constructors

StreamInitInfo

Fields

version :: Version

entitySorting :: EntitySorting

numEntities :: Maybe Word64

rootCausalHash :: Hash32

rootBranchRef :: Maybe BranchRef

data SyncError

Constructors

SyncErrorExpectedResultNotInMain CausalHash

SyncErrorDeserializationFailure DeserialiseFailure

SyncErrorMissingInitialChunk

SyncErrorMisplacedInitialChunk

SyncErrorStreamFailure Text

SyncErrorUnsupportedVersion Version

data DownloadEntitiesError

Constructors

DownloadEntitiesNoReadPermission BranchRef

DownloadEntitiesInvalidBranchRef Text BranchRef
msg, branchRef

DownloadEntitiesUserNotFound Text
userHandle

DownloadEntitiesProjectNotFound Text
project shorthand

DownloadEntitiesEntityValidationFailure EntityValidationError

data CausalDependenciesRequest

Constructors

CausalDependenciesRequest

Fields

branchRef :: BranchRef

rootCausal :: HashJWT

data CausalDependenciesChunk

A chunk of the download entities response stream.

Constructors

CausalHashDepC

Fields

causalHash :: Hash32

dependencyType :: DependencyType

data DependencyType

Constructors

CausalSpineDependency

LibDependency

newtype CBORBytes t

Wrapper for CBOR data that has already been serialized. In our case, we use this because we may load pre-serialized CBOR directly from the database, but it's also useful in allowing us to more quickly seek through a CBOR stream, since we only need to decode the CBOR when/if we actually need to use it, and can skip past it using a byte offset otherwise.

The t phantom type is the type of the data encoded in the bytestring.

Constructors

CBORBytes ByteString

newtype CBORStream a

Wrapper for a stream of CBOR data. Each chunk may not be a complete CBOR value, but the concatenation of all the chunks is a valid CBOR stream.

Constructors

CBORStream ByteString

data EntityKind

Constructors

CausalEntity

NamespaceEntity

TermEntity

TypeEntity

PatchEntity

serialiseCBORBytes :: Serialise t => t -> CBORBytes t

deserialiseOrFailCBORBytes :: Serialise t => CBORBytes t -> Either DeserialiseFailure t

Deserialize a CBORBytes value into its tagged type, throwing an error if the deserialization fails.

newtype BranchRef

Constructors

BranchRef

Fields

unBranchRef :: Text

data PullError

An error occurred while pulling code from Unison Share.

Constructors

PullError'DownloadEntities DownloadEntitiesError

PullError'Sync SyncError

data EntitySorting

Constructors

DependenciesFirst

Unsorted

newtype Version

Constructors

Version Word16