Unison.Util.Websockets

Documentation

withQueues :: forall i o m a. (MonadUnliftIO m, WebSocketsData i, WebSocketsData o) => Int -> Int -> Connection -> (Queues i o -> m a) -> m (Either ConnectionException (a, [o]))

data Queues i o

Allows interfacing with a websocket as a pair of bounded queues.

Constructors

Queues

Fields

receive :: STM (Maybe o)

send :: i -> STM Bool

data MsgOrError err a

Type used for websocket messages that can either be a message or an error.

Constructors

Msg !a

UserErr !err

DeserialiseFailure !Text

withCodeserverWebsocket :: forall m i o r e. (MonadUnliftIO m, WebSocketsData i, WebSocketsData o) => Int -> CodeserverURI -> (CodeserverId -> IO (Either e Text)) -> String -> (Queues i o -> m r) -> m (Either ConnectionException (r, [o]))

Connect a websocket to the codeserver at the given URI. The action will be called with a Queues to send and receive messages, when the action completes, the websocket connection will be closed.