Unison.Server.CodebaseServer

Documentation

data HTML

Constructors

HTML

newtype RawHtml

Constructors

RawHtml

Fields

unRaw :: ByteString

type OpenApiJSON = "openapi.json" :> Get '[JSON] OpenApi

type UnisonAndDocsAPI = UnisonLocalAPI :<|> (OpenApiJSON :<|> Raw)

type UnisonLocalAPI = ("projects" :> ProjectsAPI) :<|> ("ucm" :> UCMAPI)

type CodebaseServerAPI = NamespaceListingAPI :<|> (NamespaceDetailsAPI :<|> (DefinitionsAPI :<|> (FuzzyFindAPI :<|> (TermSummaryAPI :<|> TypeSummaryAPI))))

type ProjectsAPI = ListProjectsEndpoint :<|> (Capture "project-name" ProjectName :> (("branches" :> (ListProjectBranchesEndpoint :<|> (Capture "branch-name" ProjectBranchName :> CodebaseServerAPI))) :<|> ("diff" :> (("terms" :> ProjectDiffTermsEndpoint) :<|> ("types" :> ProjectDiffTypesEndpoint)))))

type ProjectDiffTermsEndpoint = RequiredQueryParam "oldBranchRef" ProjectBranchName :> (RequiredQueryParam "newBranchRef" ProjectBranchName :> (RequiredQueryParam "oldTerm" Name :> (RequiredQueryParam "newTerm" Name :> Get '[JSON] TermDiffResponse)))

type ProjectDiffTypesEndpoint = RequiredQueryParam "oldBranchRef" ProjectBranchName :> (RequiredQueryParam "newBranchRef" ProjectBranchName :> (RequiredQueryParam "oldType" Name :> (RequiredQueryParam "newType" Name :> Get '[JSON] TypeDiffResponse)))

type WebUI = CaptureAll "route" Text :> Get '[HTML] RawHtml

type ServerAPI = ("ui" :> WebUI) :<|> ("api" :> UnisonAndDocsAPI)

type StaticAPI = "static" :> Raw

type Authed api = Capture "token" Text :> api

type AppAPI = StaticAPI :<|> Authed ServerAPI

data BaseUrl

Constructors

BaseUrl

Fields

urlHost :: String

urlToken :: ByteString

urlPort :: Port

data DefinitionReference

Constructors

TermReference (HashQualified Name)

TypeReference (HashQualified Name)

AbilityConstructorReference (HashQualified Name)

DataConstructorReference (HashQualified Name)

data Service

Constructors

ProjectBranchUI (ProjectAndBranch ProjectName ProjectBranchName) Absolute (Maybe DefinitionReference)

Api

data URISegment

Constructors

EscapeMe Text

DontEscape Text

urlFor :: Service -> BaseUrl -> Text

Create a Service URL, either for the UI or the API

Examples:

>>> urlFor Api (BaseUrl{ urlHost = "http://localhost", urlToken = "asdf", urlPort = 1234 })
"http://localhost:1234/asdf/api"

Loose code with definition but no perspective >>> import qualified Unison.Syntax.Name as Name >>> let service = LooseCodeUI (Path.absoluteEmpty) (Just (TermReference (NameOnly (Name.unsafeFromText "base.data.List.map")))) >>> let baseUrl = (BaseUrl{ urlHost = "http://localhost", urlToken = "asdf", urlPort = 1234 }) >>> urlFor service baseUrl "http://localhost:1234/asdf/ui/non-project-code/latest/terms/base/data/List/map"

Loose code with definition and perspective >>> import qualified Unison.Syntax.Name as Name >>> let service = LooseCodeUI (Path.Absolute (Path.fromText "base.data")) (Just (TermReference (NameOnly (Name.unsafeFromText "List.map")))) >>> let baseUrl = (BaseUrl{ urlHost = "http://localhost", urlToken = "asdf", urlPort = 1234 }) >>> urlFor service baseUrl "http://localhost:1234/asdf/ui/non-project-code/latest/namespaces/base/data/;/terms/List/map"

Project with definition but no perspective >>> import qualified Unison.Syntax.Name as Name >>> import Unison.Core.Project (ProjectName (..), ProjectBranchName (..), ProjectAndBranch (..)) >>> let service = ProjectBranchUI (ProjectAndBranch (UnsafeProjectName "base") (UnsafeProjectBranchName "main")) (Path.empty) (Just (TermReference (NameOnly (Name.unsafeFromText "List.map")))) >>> let baseUrl = (BaseUrl{ urlHost = "http://localhost", urlToken = "asdf", urlPort = 1234 }) >>> urlFor service baseUrl "http://localhost:1234/asdf/ui/projects/base/main/latest/terms/List/map"

Project with definition but no perspective, contributor branch >>> import qualified Unison.Syntax.Name as Name >>> import Unison.Core.Project (ProjectName (..), ProjectBranchName (..), ProjectAndBranch (..)) >>> let service = ProjectBranchUI (ProjectAndBranch (UnsafeProjectName "unison/base") (UnsafeProjectBranchName "runarorama/contribution")) (Path.empty) (Just (TermReference (NameOnly (Name.unsafeFromText "List.map")))) >>> let baseUrl = (BaseUrl{ urlHost = "http://localhost", urlToken = "asdf", urlPort = 1234 }) >>> urlFor service baseUrl "http://localhost:1234/asdf/ui/projects/@unison/base/@runarorama/contribution/latest/terms/List/map"

Project with definition and perspective >>> import qualified Unison.Syntax.Name as Name >>> import Unison.Core.Project (ProjectName (..), ProjectBranchName (..), ProjectAndBranch (..)) >>> let service = ProjectBranchUI (ProjectAndBranch (UnsafeProjectName "unison/base") (UnsafeProjectBranchName "runarorama/contribution")) (Path.fromList ["data"]) (Just (TermReference (NameOnly (Name.unsafeFromText "List.map")))) >>> let baseUrl = (BaseUrl{ urlHost = "http://localhost", urlToken = "asdf", urlPort = 1234 }) >>> urlFor service baseUrl "http://localhost:1234/asdf/ui/projects/@unison/base/@runarorama/contribution/latest/namespaces/data/;/terms/List/map"

handleAuth :: ByteString -> Text -> Handler ()

openAPI :: OpenApi

infoObject :: Info

docsBS :: ByteString

unisonAndDocsAPI :: Proxy UnisonAndDocsAPI

api :: Proxy UnisonLocalAPI

serverAPI :: Proxy ServerAPI

appAPI :: Proxy AppAPI

app :: BackendEnv -> Runtime Symbol -> Codebase IO Symbol Ann -> FilePath -> ByteString -> Maybe String -> Application

genToken :: IO ByteString

The Token is used to help prevent multiple users on a machine gain access to each others codebases.

Generate a cryptographically secure random token. https://neilmadden.blog/2018/08/30/moving-away-from-uuids/

E.g. >>> genToken "uxf85C7Y0B6om47"

data Waiter a

Constructors

Waiter

Fields

notify :: a -> IO ()

waitFor :: IO a

mkWaiter :: IO (Waiter a)

ucmUIVar :: String

ucmPortVar :: String

ucmHostVar :: String

ucmAllowCorsHost :: String

ucmTokenVar :: String

data CodebaseServerOpts

Constructors

CodebaseServerOpts

Fields

token :: Maybe String

host :: Maybe String

port :: Maybe Int

allowCorsHost :: Maybe String

codebaseUIPath :: Maybe FilePath

defaultCodebaseServerOpts :: CodebaseServerOpts

startServer :: BackendEnv -> CodebaseServerOpts -> Runtime Symbol -> Codebase IO Symbol Ann -> (BaseUrl -> IO a) -> IO a

serveIndex :: FilePath -> Handler RawHtml

serveUI :: FilePath -> Server WebUI

corsPolicy :: Maybe String -> Middleware

server :: BackendEnv -> Runtime Symbol -> Codebase IO Symbol Ann -> FilePath -> ByteString -> Server AppAPI

serveUnisonAndDocs :: BackendEnv -> Runtime Symbol -> Codebase IO Symbol Ann -> Server UnisonAndDocsAPI

serveDocs :: Application

serveOpenAPI :: Handler OpenApi

hoistWithAuth :: forall api. HasServer api '[] => Proxy api -> ByteString -> ServerT api Handler -> ServerT (Authed api) Handler

serveProjectsCodebaseServerAPI :: Codebase IO Symbol Ann -> Runtime Symbol -> ProjectName -> ProjectBranchName -> ServerT CodebaseServerAPI (Backend IO)

resolveProjectRoot :: Codebase IO v a -> ProjectAndBranch ProjectName ProjectBranchName -> Backend IO (CausalBranch Transaction)

resolveProjectRootHash :: Codebase IO v a -> ProjectAndBranch ProjectName ProjectBranchName -> Backend IO CausalHash

serveProjectDiffTermsEndpoint :: Codebase IO Symbol Ann -> Runtime Symbol -> ProjectName -> ProjectBranchName -> ProjectBranchName -> Name -> Name -> Backend IO TermDiffResponse

contextForProjectBranch :: Codebase IO v a -> ProjectName -> ProjectBranchName -> Backend IO (PrettyPrintEnvDecl, NameSearch Transaction)

serveProjectDiffTypesEndpoint :: Codebase IO Symbol Ann -> Runtime Symbol -> ProjectName -> ProjectBranchName -> ProjectBranchName -> Name -> Name -> Backend IO TypeDiffResponse

serveProjectsAPI :: Codebase IO Symbol Ann -> Runtime Symbol -> ServerT ProjectsAPI (Backend IO)

serveUnisonLocal :: BackendEnv -> Codebase IO Symbol Ann -> Runtime Symbol -> Server UnisonLocalAPI

backendHandler :: BackendEnv -> Backend IO a -> Handler a

Orphan instances

ToSample Char

Instance details

Methods

toSamples :: Proxy Char -> [(Text, Char)]

ToParam (QueryParam' mods "newBranchRef" a :: Type)

Instance details

Methods

toParam :: Proxy (QueryParam' mods "newBranchRef" a) -> DocQueryParam

ToParam (QueryParam' mods "newTerm" a :: Type)

Instance details

Methods

toParam :: Proxy (QueryParam' mods "newTerm" a) -> DocQueryParam

ToParam (QueryParam' mods "newType" a :: Type)

Instance details

Methods

toParam :: Proxy (QueryParam' mods "newType" a) -> DocQueryParam

ToParam (QueryParam' mods "oldBranchRef" a :: Type)

Instance details

Methods

toParam :: Proxy (QueryParam' mods "oldBranchRef" a) -> DocQueryParam

ToParam (QueryParam' mods "oldTerm" a :: Type)

Instance details

Methods

toParam :: Proxy (QueryParam' mods "oldTerm" a) -> DocQueryParam

ToParam (QueryParam' mods "oldType" a :: Type)

Instance details

Methods

toParam :: Proxy (QueryParam' mods "oldType" a) -> DocQueryParam