Unison.Sync.Types

Misc. types

newtype Base64Bytes

A newtype for JSON encoding binary data.

Constructors

Base64Bytes ByteString

newtype RepoInfo

Constructors

RepoInfo

Fields

unRepoInfo :: Text

data Path

Constructors

Path

Fields

pathSegments :: NonEmpty Text

pathRepoInfo :: Path -> RepoInfo

Convert a path like arya.public.mystuff to a "repo info" by treating the first segment as a user handle.

pathCodebasePath :: Path -> [Text]

Entity types

data Entity text noSyncHash hash

Constructors

TC (TermComponent text hash)

DC (DeclComponent text hash)

P (Patch text noSyncHash hash)

PD (PatchDiff text noSyncHash hash)

N (Namespace text hash)

ND (NamespaceDiff text hash)

C (Causal hash)

data TermComponent text hash

Constructors

TermComponent [(LocalIds text hash, ByteString)]

data DeclComponent text hash

Constructors

DeclComponent [(LocalIds text hash, ByteString)]

data Patch text oldHash newHash

Constructors

Patch

Fields

textLookup :: [text]

oldHashLookup :: [oldHash]

newHashLookup :: [newHash]

bytes :: ByteString

data PatchDiff text oldHash hash

Constructors

PatchDiff

Fields

parent :: hash

textLookup :: [text]

oldHashLookup :: [oldHash]

newHashLookup :: [hash]

bytes :: ByteString

data Namespace text hash

Constructors

Namespace

Fields

textLookup :: [text]

defnLookup :: [hash]

patchLookup :: [hash]

childLookup :: [(hash, hash)]

bytes :: LocalBranchBytes

data NamespaceDiff text hash

Constructors

NamespaceDiff

Fields

parent :: hash

textLookup :: [text]

defnLookup :: [hash]

patchLookup :: [hash]

childLookup :: [(hash, hash)]

bytes :: LocalBranchBytes

data Causal hash

Constructors

Causal

Fields

namespaceHash :: hash

parents :: Set hash

data LocalIds text hash

Constructors

LocalIds

Fields

texts :: [text]

hashes :: [hash]

entityDependencies :: Ord hash => Entity text noSyncHash hash -> Set hash

Get the direct dependencies of an entity (which are actually sync'd).

FIXME use generic-lens here? (typed @hash)

data EntityType

Constructors

TermComponentType

DeclComponentType

PatchType

PatchDiffType

NamespaceType

NamespaceDiffType

CausalType

Entity Traversals

entityHashes_ :: (Applicative m, Ord hash') => (hash -> m hash') -> Entity text noSyncHash hash -> m (Entity text noSyncHash hash')

patchOldHashes_ :: Applicative m => (oldHash -> m oldHash') -> Patch text oldHash newHash -> m (Patch text oldHash' newHash)

patchNewHashes_ :: Applicative m => (newHash -> m newHash') -> Patch text oldHash newHash -> m (Patch text oldHash newHash')

patchDiffHashes_ :: Applicative m => (hash -> m hash') -> PatchDiff text noSyncHash hash -> m (PatchDiff text noSyncHash hash')

namespaceDiffHashes_ :: Applicative m => (hash -> m hash') -> NamespaceDiff text hash -> m (NamespaceDiff text hash')

causalHashes_ :: (Applicative m, Ord hash') => (hash -> m hash') -> Causal hash -> m (Causal hash')

Request/response types

Get causal hash by path

newtype GetCausalHashByPathRequest

Constructors

GetCausalHashByPathRequest

Fields

path :: Path

data GetCausalHashByPathResponse

Constructors

GetCausalHashByPathSuccess (Maybe HashJWT)

GetCausalHashByPathNoReadPermission Path

GetCausalHashByPathUserNotFound

GetCausalHashByPathInvalidRepoInfo Text RepoInfo

Download entities

data DownloadEntitiesRequest

Constructors

DownloadEntitiesRequest

Fields

repoInfo :: RepoInfo

hashes :: NESet HashJWT

data DownloadEntitiesResponse

Constructors

DownloadEntitiesSuccess (NEMap Hash32 (Entity Text Hash32 HashJWT))

DownloadEntitiesFailure DownloadEntitiesError

data DownloadEntitiesError

Constructors

DownloadEntitiesNoReadPermission RepoInfo

DownloadEntitiesInvalidRepoInfo Text RepoInfo
msg, repoInfo

DownloadEntitiesUserNotFound Text
userHandle

DownloadEntitiesProjectNotFound Text
project shorthand

DownloadEntitiesEntityValidationFailure EntityValidationError

Upload entities

data UploadEntitiesRequest

Constructors

UploadEntitiesRequest

Fields

repoInfo :: RepoInfo

entities :: NEMap Hash32 (Entity Text Hash32 Hash32)

data UploadEntitiesResponse

Constructors

UploadEntitiesSuccess

UploadEntitiesFailure UploadEntitiesError

data UploadEntitiesError

Constructors

UploadEntitiesError'EntityValidationFailure EntityValidationError

UploadEntitiesError'HashMismatchForEntity HashMismatchForEntity

UploadEntitiesError'InvalidRepoInfo Text RepoInfo
msg, repoInfo

UploadEntitiesError'NeedDependencies (NeedDependencies Hash32)

UploadEntitiesError'NoWritePermission RepoInfo

UploadEntitiesError'ProjectNotFound Text
project shorthand

UploadEntitiesError'UserNotFound Text
userHandle

Common/shared error types

data HashMismatchForEntity

Constructors

HashMismatchForEntity

Fields

supplied :: Hash32

computed :: Hash32

data InvalidParentage

Constructors

InvalidParentage

Fields

parent :: Hash32

child :: Hash32

data NeedDependencies hash

Constructors

NeedDependencies

Fields

missingDependencies :: NESet hash

data EntityValidationError

The ways in which validating an entity may fail.

Constructors

EntityHashMismatch EntityType HashMismatchForEntity

UnsupportedEntityType Hash32 EntityType

InvalidByteEncoding Hash32 EntityType Text

HashResolutionFailure Hash32