Unison.Util.Servant.CBOR

Description

Servant configuration for the CBOR media type

Adapted from https://hackage.haskell.org/package/servant-serialization-0.3/docs/Servant-API-ContentTypes-SerialiseCBOR.html via MIT license

Documentation

data CBOR

Content-type for encoding and decoding objects as their CBOR representations

type UnknownCBORBytes = CBORBytes Unknown

newtype CBORBytes t

Wrapper for CBOR data that has already been serialized. In our case, we use this because we may load pre-serialized CBOR directly from the database, but it's also useful in allowing us to more quickly seek through a CBOR stream, since we only need to decode the CBOR when/if we actually need to use it, and can skip past it using a byte offset otherwise.

The t phantom type is the type of the data encoded in the bytestring.

Constructors

CBORBytes ByteString

newtype CBORStream a

Wrapper for a stream of CBOR data. Each chunk may not be a complete CBOR value, but the concatenation of all the chunks is a valid CBOR stream.

Constructors

CBORStream ByteString

deserialiseOrFailCBORBytes :: Serialise t => CBORBytes t -> Either DeserialiseFailure t

Deserialize a CBORBytes value into its tagged type, throwing an error if the deserialization fails.

serialiseCBORBytes :: Serialise t => t -> CBORBytes t

decodeCBORBytes :: Serialise t => CBORBytes t -> Decoder s t

decodeUnknownCBORBytes :: Serialise t => UnknownCBORBytes -> Decoder s t

serialiseUnknownCBORBytes :: Serialise t => t -> UnknownCBORBytes